requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
pgvector>=0.2.0

# Document Parsing
python-docx>=1.1.0
python-pptx>=0.6.23
openpyxl>=3.1.2
pdfplumber>=0.10.3
PyMuPDF>=1.23.0
pandas>=2.0.0

# Vector Embeddings
sentence-transformers>=2.2.0
numpy>=1.24.0

# LLM Integration
openai>=1.10.0
anthropic>=0.18.0
boto3>=1.34.0

# File Storage
boto3>=1.34.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0
aiofiles>=23.2.1

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Development
black>=24.1.0
isort>=5.13.0
mypy>=1.8.0
